changed the shares type from double to long to remove the fucking problem of decimal and updated all realted classes of it

+++ src/main/java/majorproject/maf/dto/SellRequest.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class SellRequest {
-    private double Quantity;
+    private long Quantity;
     private String symbol;
 }
